requirements.txt
# Core Dependencies
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1

# ML Libraries
xgboost==2.1.1
lightgbm==4.5.0
catboost==1.2.7
optuna==4.2.0

# Visualization
plotly==5.24.1
matplotlib==3.10.0
seaborn==0.13.2

# Web Interface
streamlit==1.39.0

# Data Handling
openpyxl==3.1.5 # Excel support
xlsxwriter==3.2.0 # Excel writing
python-docx==1.1.2 # Word document support

# Statistical Analysis
statsmodels==0.14.4
shap==0.46.0
lime==Available

# System & Utilities
psutil==6.0.0 # System monitoring
joblib==1.4.2 # Model persistence
tqdm==4.66.5 # Progress bars

# Development Tools
jupyter==1.1.1
ipywidgets==8.1.5

# Testing & Quality
pytest==8.3.3
pylint==3.3.1
mypy==1.11.2
black==24.10.0
flake8==7.1.1

# Code Formatting & Analysis
autopep8==2.3.2
yapf==0.40.2
rope==1.13.0

# Type Checking
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2

# Optional Dependencies
pdpbox==Available # Partial dependence plots
itables==2.2.2 # Interactive tables
